+++ libcef_dll/ctocpp/frame_ctocpp.h
@@ -54,7 +54,7 @@ public:
   virtual bool IsMain() OVERRIDE;
   virtual bool IsFocused() OVERRIDE;
   virtual CefString GetName() OVERRIDE;
-  virtual long long GetIdentifier() OVERRIDE;
+  virtual int64 GetIdentifier() OVERRIDE;
   virtual CefRefPtr<CefFrame> GetParent() OVERRIDE;
   virtual CefString GetURL() OVERRIDE;
   virtual CefRefPtr<CefBrowser> GetBrowser() OVERRIDE;
